Configurando archivo POM, configuracion de dependencia y versiones para el correcto funcionamiento de Swagger

+++ resena/src/main/java/com/proyectofs1/resena/service/ResenaService.java
@@ -71,10 +71,14 @@ public class ResenaService {
     }
 
     public void deleteById(Long id) {
-        log.info("Ejecutando eliminación de la reseña con ID: {}", id);
-        resenaRepository.deleteById(id);
-        log.info("Reseña con ID: {} eliminada correctamente.", id);
+    log.info("Ejecutando eliminación de la reseña con ID: {}", id);
+    if (!resenaRepository.existsById(id)) {
+        log.warn("Fallo al eliminar: No existe reseña con ID {}", id);
+        throw new IllegalArgumentException("No se encontró la reseña a eliminar.");
     }
+    resenaRepository.deleteById(id);
+    log.info("Reseña con ID: {} eliminada correctamente.", id);
+}
 
     // --- Métodos de Conversión ---
